<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- User server mode for GC
- Remove SecurityType cache type check
#### Cloud Performance Tests Py
##### MASTER BasicTemplateBenchmark
28.38 seconds at 93k data points per second. Processing total of 2,639,319 data points.
27.50 seconds at 96k data points per second. Processing total of 2,639,319 data points.

##### PR BasicTemplateBenchmark
24.48 seconds at 108k data points per second. Processing total of 2,639,319 data points.
25.13 seconds at 105k data points per second. Processing total of 2,639,319 data points.

##### MASTER EmptyMinute400EquityBenchmark
30.58 seconds at 275k data points per second. Processing total of 8,401,108 data points.
28.97 seconds at 290k data points per second. Processing total of 8,401,108 data points.

##### PR EmptyMinute400EquityBenchmark
12.42 seconds at 676k data points per second. Processing total of 8,401,108 data points.
13.07 seconds at 643k data points per second. Processing total of 8,401,108 data points.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
N/A
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Need for speed
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Running and profiling algorithms
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->